+++ R/multi.ace_fun.R
@@ -196,12 +196,16 @@ add.state.names <- function(estimation, character_states, tree) {
 }
 
 ## Translating the likelihood table into a vector of characters
-translate.likelihood <- function(character, threshold, select.states, special.tokens) {
+translate.likelihood <- function(character, threshold, select.states, special.tokens, do_sample) {
     ## Translate the likelihood table
-    threshold.fun <- function(taxon, threshold, select.states, special.tokens) {
-        return(paste(select.states(taxon, threshold), collapse = sub("\\\\", "", special.tokens["uncertainty"])))
+    threshold.fun <- function(taxon, threshold, select.states, special.tokens, do_sample) {
+        if(!do_sample) {
+            return(paste(select.states(taxon, threshold), collapse = sub("\\\\", "", special.tokens["uncertainty"])))
+        } else {
+            return(select.states(taxon, threshold))
+        }
     }
-    translated_states <- apply(character, 1, threshold.fun, threshold, select.states, special.tokens)
+    translated_states <- apply(character, 1, threshold.fun, threshold, select.states, special.tokens, do_sample)
     ## Replace empty states by uncertainties
     replace.empty.states <- function(x, character, special.tokens) {
         return(ifelse(x == "", paste(colnames(character), collapse = sub("\\\\", "", special.tokens["uncertainty"])), x))
@@ -212,7 +216,7 @@ translate.likelihood <- function(character, threshold, select.states, special.to
 }
 
 ## Function for running ace on a single tree.
-one.tree.ace <- function(args_list, special.tokens, invariants, characters_states, threshold.type, threshold, invariant_characters_states, verbose) {
+one.tree.ace <- function(args_list, special.tokens, invariants, characters_states, threshold.type, threshold, invariant_characters_states, verbose, do_sample) {
 
     if(verbose) body(castor.ace)[[2]] <- substitute(cat("."))
     # if(verbose) cat("Running ancestral states estimations:\n")
@@ -249,29 +253,45 @@ one.tree.ace <- function(args_list, special.tokens, invariants, characters_state
                                               } else {
                                                 return(names(taxon[taxon >= threshold]))
                                               }
-                                             }})
+                                             }},
+        sample  = {select.states <- function(taxon, threshold){
+                                             if(all(is.na(taxon))) {
+                                                return(NA)
+                                             } else {
+                                                return(sample(names(taxon), size = threshold, prob = taxon, replace = TRUE))
+                                             }}}
+                                         )
     
     ## Estimate the ancestral states
-    ancestral_states <- lapply(ancestral_estimations, translate.likelihood, threshold, select.states, special.tokens)
+    ancestral_states <- lapply(ancestral_estimations, translate.likelihood, threshold, select.states, special.tokens, do_sample)
 
     ## Add invariants characters back in place
     if(length(invariants) > 0) {
         ## Replicate the invariant characters
-        invariant_ancestral <- lapply(invariant_characters_states, function(x, n) rep(ifelse(length(x == 0), special.tokens["missing"], x), n), args_list[[1]]$tree$Nnode)
+        invariant_ancestral <- lapply(invariant_characters_states, function(x, n) rep(ifelse(length(x) == 0, special.tokens["missing"], x), n),n = args_list[[1]]$tree$Nnode)
 
         ## Combine the final dataset
         output <- replicate(length(args_list)+length(invariants), list())
         ## Fill the final dataset
-        output[invariants] <- invariant_ancestral
+        if(!do_sample) {
+            output[invariants] <- invariant_ancestral
+        } else {
+            output[invariants] <- list(matrix(invariant_ancestral[[1]][1], ncol = length(invariant_ancestral[[1]]), nrow = threshold))
+        }
         output[-invariants] <- ancestral_states
         ancestral_states <- output
     }
 
     ## Replace NAs
-    replace.NA <- function(character, characters_states, special.tokens) {
-        return(sapply(character, function(x) ifelse(x[[1]] == "NA", paste0(characters_states, collapse = sub("\\\\", "", special.tokens["uncertainty"])), x)))
+    replace.NA <- function(character, characters_states, special.tokens, do_sample) {
+        na_removed <- sapply(character, function(x) ifelse(x[[1]] == "NA", paste0(characters_states, collapse = sub("\\\\", "", special.tokens["uncertainty"])), x))
+        if(do_sample) {
+            return(matrix(na_removed, nrow = dim(character)[1], ncol = dim(character)[2], byrow = TRUE, dimnames = dimnames(character)))
+        } else {
+            return(na_removed)
+        }
     }
-    ancestral_states[-invariants] <- mapply(replace.NA, ancestral_states[-invariants], characters_states, MoreArgs = list(special.tokens = special.tokens), SIMPLIFY = FALSE)
+    ancestral_states[-invariants] <- mapply(replace.NA, ancestral_states[-invariants], characters_states, MoreArgs = list(special.tokens = special.tokens, do_sample = do_sample), SIMPLIFY = FALSE)
 
     ## Sort the details list
     if(!is.null(args_list[[1]]$details)) {
@@ -291,20 +311,38 @@ one.tree.ace <- function(args_list, special.tokens, invariants, characters_state
 }
 
 ## Bind the continuous and discrete characters and reorder them
-bind.characters <- function(continuous, discrete, order) {
-    bound <- cbind(as.data.frame(continuous), as.data.frame(discrete))
+bind.characters <- function(continuous, discrete, order, do_sample) {
+
+    if(!do_sample) {
+        continuous <- list(continuous)
+        discrete <- list(discrete)
+    }
+
+    ## Bind the lists
+    bound <- mapply(function(x, y) cbind(as.data.frame(x), as.data.frame(y)), continuous, discrete, SIMPLIFY = FALSE)
     ## Get the new character IDs
-    cont_names <- colnames(bound)[1:ncol(continuous)]
-    disc_names <- colnames(bound)[-c(1:ncol(continuous))]
+    cont_names <- colnames(bound[[1]])[1:ncol(continuous[[1]])]
+    disc_names <- colnames(bound[[1]])[-c(1:ncol(continuous[[1]]))]
+
     ## Rename discrete if they have the names in common
     if(any(disc_names %in% cont_names)) {
         disc_names <- paste0("c", disc_names)
-        colnames(bound)[-c(1:ncol(continuous))] <- disc_names
+        bound <- lapply(bound, function(x, n_cont, disc_names) {colnames(x)[-c(1:n_cont)] <- disc_names; return(x)}, n_cont = ncol(continuous[[1]]), disc_names = disc_names)
     }
+
     ## Reorder the characters to match the input order
-    ordering <- matrix(c(1:ncol(bound), c(order$continuous, order$discrete)), ncol = 2, byrow = FALSE, dimnames = list(c(cont_names, disc_names), c("out", "in")))
-    return(bound[, names(sort(ordering[, 2, drop = TRUE]))])
+    ordering <- matrix(c(1:ncol(bound[[1]]), c(order$continuous, order$discrete)), ncol = 2, byrow = FALSE, dimnames = list(c(cont_names, disc_names), c("out", "in")))
+    ordered_characters <- lapply(bound, function(x, ordering) x[, names(sort(ordering[, 2, drop = TRUE]))], ordering = ordering)
+
+    ## Output
+    if(!do_sample) {
+        return(ordered_characters[[1]])
+    } else {
+        return(ordered_characters)
+    }
 }
+
+
 ## Bind the continuous and discrete details and reorder them
 bind.details <- function(continuous, discrete, order) {
     ## Reorder the details out per characters
@@ -335,3 +373,40 @@ make.list <- function(results) {
     return(unlist(apply(results, 1, list), recursive = FALSE))
 }
 ## Make the dispRity object out of the results (only for continuous)
+
+
+
+
+
+# Samples a single estimate from an ace output
+#' @param ace the ace output
+#' @param sample.fun a list with two elements: fun the sampling function and param a named list of parameters and how to sample them
+#' @param samples the number of samples (default is 1)
+sample.ace <- function(ace, sample.fun, samples = 1) {
+    ## Generate the parameters list
+    make.param <- function(one_node, sample.fun, samples) {
+        params <- lapply(sample.fun$param, function(fun, data) return(fun(data)), data = one_node)
+        return(list(fun = sample.fun$fun, param = c(n = samples, params)))
+    }
+    fun_list <- apply(ace$CI95, 1, make.param, sample.fun, samples)
+
+    ## Sample all the values
+    return(lapply(fun_list, function(x) do.call(what = x$fun, args = x$param)))
+}
+
+## Testing the sample.fun argument. returns FALSE if fail, TRUE if OK
+test.sample.fun <- function(one_sample, n = 1) {
+    ## Sample the parameters
+    test <- try(params <- lapply(one_sample$param, function(x) do.call(x, args = list(c(1,2,3)))), silent = TRUE)
+    if(is(test, "try-error")) {
+        return(FALSE)
+    }
+    ## Add n
+    params$n <- n
+    ## Test the function
+    test <- try(do.call(one_sample$fun, args = params), silent = TRUE)
+    if(is(test, "try-error")) {
+        return(FALSE)
+    }
+    return(TRUE)
+}