+++ tests/testthat/test-multi.ace.R
@@ -1,4 +1,3 @@
-#context("multi.ace")
 test_that("model internals works", {
 
     ## with corStruc
@@ -15,7 +14,6 @@ test_that("model internals works", {
     expect_equal(check.model.class(one_model = 1, available_models = available_models_continuous), "numeric")
 })
 
-## Test
 test_that("multi.ace works", {
 
     ## Sanitizing works
@@ -420,6 +418,12 @@ test_that("multi.ace works", {
     # expect_equal(rownames(results$details[[1]]$transition_matrix[[9]]), c("0","1","2"))
     # expect_is(results$details[[2]]$loglikelihood[[1]], "numeric")
 
+
+    ## No match check
+    tree <- makeNodeLabel(rcoal(10))
+    matrix_continuous <- space.maker(elements = 9, dimensions = 3, distribution = rnorm, elements.names = tree$node.label)
+    expect_warning(error <- capture_error(multi.ace(matrix_continuous, tree)))
+    expect_equal(error[[1]], "No match between the tips in the tree and the rows in the data.")
 })
 
 test_that("multi.ace works with continuous and mix", {
@@ -472,4 +476,116 @@ test_that("multi.ace works with continuous and mix", {
     expect_warning(test <- multi.ace(data = data, tree = tree, threshold = 0.75))
     expect_is(test, "data.frame")
     expect_equal(dim(test), c(14,9))
+})
+
+test_that("sample.ace works", {
+    ## Create a quick ace
+    set.seed(1)
+    ## The tree
+    tree <- rcoal(15)
+    tree <- makeNodeLabel(tree)
+    ## The matrix
+    data <- space.maker(elements = 15, dimensions = 5, distribution = rnorm, elements.name = tree$tip.label)
+    ## The sampling function
+    sample.fun <- list(fun = runif, param = list(min = min, max = max))
+
+    expect_warning(ace <- ape::ace(x = data[,1], phy = tree))
+    
+    ## Sample the ace output
+    set.seed(1)
+    out <- sample.ace(ace, sample.fun, samples = 2)
+    expect_equal(length(out), Nnode(tree))
+    expect_equal(length(out[[1]]), 2)
+    ## Sampling works
+    for(i in 1:Nnode(tree)) {
+        expect_true(all(out[[i]] > ace$CI95[i,1]))
+        expect_true(all(out[[i]] < ace$CI95[i,2]))
+    }
+})
+
+test_that("multi.ace works with sample", {
+    set.seed(1)
+    ## The tree
+    tree <- rcoal(10)
+    tree <- makeNodeLabel(tree)
+    ## The matrix
+    data <- data_continuous <- cbind(runif(10, 0, 1), runif(10, 10, 20), runif(10, 100, 200))
+    rownames(data) <- tree$tip.label
+
+    # Test with continuous
+    expect_warning(test <- multi.ace(data = data, tree = tree, sample = 2, output = "combined.matrix", verbose = FALSE))
+    expect_is(test, "list")
+    ## Correct number of samples
+    expect_equal(length(test), 2)
+    ## Correct rows
+    expect_equal(rownames(test[[1]]), c(tree$tip.label, tree$node.label))
+    ## Correct character estimates
+    expect_true(all(test[[1]][,1] < 2))
+    expect_true(all(test[[1]][,2] < 50))
+    expect_true(all(test[[1]][,2] > -5))
+    expect_true(all(test[[1]][,3] > 100))
+
+    ## test with sample.fun option
+    error <- capture_error(test <- multi.ace(data = data, tree = tree, sample = 2, sample.fun = runif, output = "combined.matrix", verbose = FALSE))
+    expect_equal(error[[1]], "sample.fun must be of class list.")
+    sample.fun <- list(fun = rnorm, param = list(max = max, min = min))
+    error <- capture_error(test <- multi.ace(data = data, tree = tree, sample = 2, sample.fun = sample.fun, output = "combined.matrix", verbose = FALSE))
+    expect_equal(error[[1]], "The sample function is not formatted correctly and cannot generate a distribution.\nCheck the ?multi.ace manual for more details.")
+    ## Works with a list of sample funs
+    sample.fun <- list(
+        list(fun = runif, param = list(max = max, min = min)),
+        list(fun = runif, param = list(max = max, min = min)),
+        list(fun = rnorm, param = list(mean = mean, sd = function(x)return(diff(range(x))/4))))
+    expect_warning(test <- multi.ace(data = data, tree = tree, sample = 2, sample.fun = sample.fun, output = "combined.matrix", verbose = FALSE))
+    expect_is(test, "list")
+    ## But doesn't if sample fun is badly formated
+    sample.fun <- list(
+        list(fun = runif, param = list(max = max, min = min)),
+        list(fun = rnorm, param = list(max = max, min = min)),
+        list(fun = runif, param = list(mean = mean, sd = function(x)return(diff(range(x))/4))))
+    error <- capture_error(test <- multi.ace(data = data, tree = tree, sample = 2, sample.fun = sample.fun, output = "combined.matrix", verbose = FALSE))
+    expect_equal(error[[1]], "The following sample functions are not formated correctly and cannot generate a distribution: 2, 3.\nCheck the ?multi.ace manual for more details.")
+
+    ## Test with discrete characters
+    set.seed(8) 
+    data <- data_discrete <- sim.morpho(tree, characters = 2, model = "ER", rates = c(rgamma, rate = 10, shape = 5), invariant = FALSE)
+    data[,2] <- data_discrete[,2] <- as.character(sample(c(1,2,3), 10, replace = TRUE))
+
+    test <- multi.ace(data = data, tree = tree, sample = 20, output = "combined.matrix", verbose = FALSE)
+    expect_is(test, "list")
+    ## Correct number of samples
+    expect_equal(length(test), 20)
+    ## Correct rows
+    expect_equal(rownames(test[[1]]), c(tree$tip.label, tree$node.label))
+    ## Correct character estimates
+    expect_true(all(as.numeric(test[[1]][,1]) %in% c(0,1)))
+    expect_true(all(as.numeric(test[[1]][,2]) %in% c(1,2,3)))
+
+    ## test with an invariant
+    data <- cbind(data, rep("0", 10))
+    expect_warning(test <- multi.ace(data = data, tree = tree, sample = 20, output = "combined.matrix", verbose = FALSE))
+    expect_is(test, "list")
+    ## Correct number of samples
+    expect_equal(length(test), 20)
+    ## Correct rows
+    expect_equal(rownames(test[[1]]), c(tree$tip.label, tree$node.label))
+    ## Correct character estimates
+    expect_true(all(as.numeric(test[[1]][,1]) %in% c(0,1)))
+    expect_true(all(as.numeric(test[[1]][,2]) %in% c(1,2,3)))
+    expect_true(all(test[[1]][,3] == "0"))
+
+    ## Test with mixed characters
+    data <- data.frame(data_discrete, data_continuous)
+    expect_warning(test <- multi.ace(data = data, tree = tree, sample = 5, output = "matrix", verbose = FALSE))
+    ## Correct number of samples
+    expect_equal(length(test), 5)
+    ## Correct rows
+    expect_equal(rownames(test[[1]]), c(tree$node.label))
+    ## Correct character estimates
+    expect_true(all(as.numeric(test[[1]][,1]) %in% c(0,1)))
+    expect_true(all(as.numeric(test[[1]][,2]) %in% c(1,2,3)))
+    expect_true(all(test[[1]][,3] < 2))
+    expect_true(all(test[[1]][,4] < 51))
+    expect_true(all(test[[1]][,4] > -5))
+    expect_true(all(test[[1]][,5] > 100))
 })